% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-params.R
\name{set_map_match_params}
\alias{set_map_match_params}
\title{Set map matching parameters}
\usage{
set_map_match_params(
  config,
  candidates = 8L,
  radius = 300,
  error = 50,
  factor = 1.5,
  vmax = 30
)
}
\arguments{
\item{config}{A configuration object.}

\item{candidates}{Number of candidates.}

\item{radius}{Search radius (unit: map unit).}

\item{error}{GPS sensor error (unit: map unit).}

\item{factor}{Scale factor to limit shortest path search. Only applicable for
stmatch.}

\item{vmax}{Maximum vehicle speed (unit: map unit). Only applicable for
stmatch.}
}
\value{
A list
}
\description{
Set map matching parameters
}
