% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_output_params}
\alias{set_output_params}
\title{Set output parameters}
\usage{
set_output_params(file, fields = "opath")
}
\arguments{
\item{file}{Output file name.}

\item{fields}{Output fields name.}
}
\value{
A list.
}
\description{
Set output parameters
}
