% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-config.R
\name{stmatch_base_config}
\alias{stmatch_base_config}
\title{Define base configuration for using the STMatch map matching algorithm}
\usage{
stmatch_base_config(
  output,
  output_fields = "all",
  log_level = 2L,
  use_omp = FALSE,
  step = 100L
)
}
\arguments{
\item{output}{Output file name.}

\item{output_fields}{Output fields name, one or more in (opath, cpath, tpath,
ogeom, mgeom, pgeom, offset, error, spdist, tp, ep, all).}

\item{log_level}{Log level (default: 2 (infor)), 0-trace, 1-debug, 2-info,
3-warn, 4-err, 5-critical, 6-off.}

\item{use_omp}{If specified, run map matching in multiple threads.}

\item{step}{Number of trajectories to report the progress of map matching.}
}
\value{
A configuration object.
}
\description{
Define base configuration for using the STMatch map matching algorithm
}
