% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_map_match_params}
\alias{set_map_match_params}
\title{Set map matching parameters}
\usage{
set_map_match_params(
  candidates = 8,
  radius = 300,
  error = 50,
  factor = 1.5,
  vmax = 30
)
}
\arguments{
\item{candidates}{Number of candidates.}

\item{radius}{Search radius.}

\item{error}{GPS sensor error.}

\item{factor}{Scale factor to limit shortest path search. Only applicable for stmatch.}

\item{vmax}{Maximum vehicle speed. Only applicable for stmatch.}
}
\description{
Set map matching parameters
}
