% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_network_params}
\alias{set_network_params}
\title{Set network parameters}
\usage{
set_network_params(
  file,
  id = "id",
  source = "source",
  target = "target",
  mode = "drive"
)
}
\arguments{
\item{file}{Network file name.}

\item{id}{Network id column name.}

\item{source}{Network source column name.}

\item{target}{Network target column name.}

\item{mode}{Network mode name. Only applicable to OpenStreetMap (OSM) network.}
}
\value{
A list.
}
\description{
Set network parameters
}
